data/requirements-etl.txt
pandas>=2.0
openpyxl>=3.1
pdfplumber>=0.11
google-api-python-client>=2.120.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
xlrd>=2.0.1
